fix(purchases): accept fractional prices in RefreshProductPricing
Counter/admin receive intermittently failed with a generic "Request
failed" when a received line's cost or selling price was fractional
(e.g. 3.4, from a whole-bill discount spread pro-rata over small items).

RefreshProductPricing's UPDATE used `CASE WHEN $1 > 0 ...`, so Postgres
inferred $1/$2 as integer from the bare `0` literal and rejected the
decimal string with "invalid input syntax for type integer" (22P02).
Whole-number prices parsed as int and passed, hence the intermittency.

Cast the parameters explicitly to numeric ($1::numeric). Syntax-only:
no schema or data change, fractions preserved exactly. Repo-wide audit
found this decimal-vs-bare-int-literal trap nowhere else. Adds a
DB-backed regression test.

+++ internal/features/purchases/purchases.go
@@ -126,10 +126,14 @@ func (r *Repository) MarkHasExpiry(ctx context.Context, productID int64) error {
 // product's real cost, which the info popup shows and the COGS zero-cost rescue
 // falls back to. Zero on either field leaves that field untouched.
 func (r *Repository) RefreshProductPricing(ctx context.Context, productID int64, cost, selling decimal.Decimal) error {
+	// $1/$2 are cast to numeric explicitly: with a bare `$1 > 0` Postgres infers
+	// the parameter type from the integer literal 0 and then rejects a fractional
+	// price like "3.4" (22P02) — the intermittent counter-receive failure. Whole
+	// prices happened to parse as int, decimals did not.
 	_, err := r.q.ExecContext(ctx, `
 		UPDATE products
-		SET cost_price    = CASE WHEN $1 > 0 THEN $1 ELSE cost_price END,
-		    selling_price = CASE WHEN $2 > 0 THEN $2 ELSE selling_price END
+		SET cost_price    = CASE WHEN $1::numeric > 0 THEN $1::numeric ELSE cost_price END,
+		    selling_price = CASE WHEN $2::numeric > 0 THEN $2::numeric ELSE selling_price END
 		WHERE id = $3`, cost, selling, productID)
 	return err
 }
